Update declared version, application keying and adjust markdown
I've updated the declared version requirement to be `~> 0.1.0` as this is the current release series. The previous 0.0.1 series doesn't seem to exist.

Also, I updated the way in which application dependencies are declared to use the `extra_applications` key to match what `mix new` seems to generate these days. 

Finally, I tweaked the markdown formatting so that these configuration examples are displayed nicely.